+++ lib/oneview-sdk/resource/api500/synergy/event.rb
@@ -1,7 +1,7 @@
-# (c) Copyright 2016-2017 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2017 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
+# You may not use this file except in compliance with the License.
 # You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software distributed
@@ -9,7 +9,14 @@
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
-# Gem version defined here
+require_relative '../../api200/event'
+
 module OneviewSDK
-  VERSION = '4.5.1'.freeze
+  module API500
+    module Synergy
+      # Event resource implementation for API500 Synergy
+      class Event < OneviewSDK::API200::Event
+      end
+    end
+  end
 end