+++ spec/integration/resource/api200/id_pool/delete_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+
+klass = OneviewSDK::IDPool
+RSpec.describe klass, integration: true, type: DELETE, sequence: rseq(klass) do
+  let(:current_client) { $client }
+  include_examples 'IDPoolDeleteExample', 'integration context'
+end