+++ spec/integration/resource/api300/c7000/id_pool/create_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+
+klass = OneviewSDK::API300::C7000::IDPool
+RSpec.describe klass, integration: true, type: CREATE, sequence: seq(klass) do
+  let(:current_client) { $client_300 }
+  include_examples 'IDPoolCreateExample', 'integration api300 context'
+end